Exit if kolla image build failed
Previously, image build failure was ignored since the failure is
caused by images that we do not need. Then, we only build what
we really need. So now, image build failure should not be ginored
any more.

This PS also do clean up task after failure occured, althrough
this is  not really necessary when running in a temporary VM. But
it is helpful when running manually.

Change-Id: Ica8771e80bc313795598e18dc0175177ccb6c11a
Signed-off-by: Zhijiang Hu <hu.zhijiang@zte.com.cn>